Fixed Recording Page structure
Formatted div's, place things where I want them to be.
